lleva registro de los videos descargados

+++ ytls.py
@@ -2,6 +2,7 @@ from youtube_dl import YoutubeDL
 import re
 import clipboard
 import json
+import os.path
 
 options = {
     'outtmpl': 'music/%(title)s-%(id)s.%(ext)s',
@@ -25,32 +26,43 @@ while url_base not in url:
     input(f'Press enter to {msg}')
 
 with YoutubeDL(options) as ydl:
-    video = ydl.extract_info(url)
-    value = video.get('description', None)
-    names = pattern.findall(value)
-    videos = []
-    for name in names:
-        search = f'ytsearch1:{name}'
-        video = ydl.extract_info(search)['entries'][0]
-        video = {key: video.get(key) for key in ['id', 'artist', 'title']}
-        print(video)
-        videos.append(video)
-    with open('videos.json', 'w') as file:
-        json.dump(videos, file, indent=2)
-    print('download...')
-    with open('videos.json', 'r') as file:
-        videos = json.load(file)
-    ydl.params['simulate'] = False
-    ydl.params['quiet'] = False
+    source = ydl.extract_info(url)
+    source_id = source.get("id")
+    file_name = f'videos-{source_id}.json'
+    if not os.path.exists(file_name):
+        try:
+            source_description = source.get('description', None)
+            names = pattern.findall(source_description)
+            videos = []
+            i = 0
+            for name in names:
+                search = f'ytsearch1:{name}'
+                vid = ydl.extract_info(search)['entries'][0]
+                vid = {key: vid.get(key) for key in ['id', 'artist', 'title']}
+                vid['downloaded'] = False
+                videos.append(vid)
+                i += 1
+                if i == 3:
+                    break
+        finally:
+            with open(file_name, 'w') as file:
+                json.dump(videos, file, indent=2)
     try:
+        with open(file_name, 'r') as file:
+            videos = json.load(file)
+        ydl.params['simulate'] = False
+        ydl.params['quiet'] = False
         for video in videos:
-            print(video)
-            url = f'{url_base}{video.get("id")}'
-            ydl.download([url])
-            video['downloaded'] = True
+            if not video.get('downloaded'):
+                print(video)
+                url = f'{url_base}{video.get("id")}'
+                ydl.download([url])
+                video['downloaded'] = True
+            else:
+                print(video, 'ready!')
     except Exception as exception:
         raise exception
     finally:
-        with open('videos.json', 'w') as file:
+        with open(file_name, 'w') as file:
             json.dump(videos, file, indent=2)
